cosas

+++ ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: dsanchez <dsanchez@student.42madrid.c      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/04 13:05:53 by dsanchez          #+#    #+#             */
-/*   Updated: 2021/09/04 13:09:41 by dsanchez         ###   ########.fr       */
+/*   Updated: 2021/09/12 20:59:13 by dsanchez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 unsigned int	ft_strlcat(char *dest, char *src, unsigned int size)
 {
-	unsigned int	i;
-	unsigned int	j;
-	unsigned int	res;
+	size_t	i;
+	size_t	j;
+	size_t	res;
 
 	i = ft_strlen(dest);
 	res = ft_strlen(src);
